feat(apps/auction_web/lib/auction_web/templates/user/show.html.eex): show bits on user page
