+++ auction_umbrella/apps/auction/lib/auction.ex
@@ -1,4 +1,5 @@
 defmodule Auction do
+  import Ecto.Query
   alias Auction.{Item, User, Password, Bid}
 
   @repo Auction.Repo
@@ -67,5 +68,18 @@ defmodule Auction do
     id
     |> get_item()
     |> @repo.preload(bids: [:user])
+  end
+
+  def get_bits_for_user(user) do
+    query =
+      from(b in Bid,
+        # The pin operator makes it so that the previously bound value is used
+        where: b.user_id == ^user.id,
+        order_by: [desc: :inserted_at],
+        preload: :item,
+        limit: 10
+      )
+
+    @repo.all(query)
   end
 end
