+++ auction_umbrella/apps/auction_web/lib/auction_web.ex
@@ -37,6 +37,7 @@ defmodule AuctionWeb do
       import Phoenix.Controller,
         only: [get_flash: 1, get_flash: 2, view_module: 1, view_template: 1]
 
+      import AuctionWeb.GlobalHelpers
       # Include shared imports and aliases for views
       unquote(view_helpers())
     end
