+++ auction_umbrella/apps/auction_web/lib/auction_web/views/global_helpers.ex
@@ -1,0 +1,15 @@
+defmodule AuctionWeb.GlobalHelpers do
+  def integer_to_currency(cents) do
+    dollar_and_cents =
+      cents
+      |> Decimal.div(100)
+      |> Decimal.round(2)
+
+    "$#{dollar_and_cents}"
+  end
+
+  def formatted_datetime(datetime) do
+    datetime
+    |> Timex.format!("{0D}/{0M}/{YYYY} {h12}:{m}")
+  end
+end
