+++ auction_umbrella/apps/auction_web/lib/auction_web/views/item_view.ex
@@ -1,11 +1,3 @@
 defmodule AuctionWeb.ItemView do
   use AuctionWeb, :view
-
-  def integer_to_currency(cents) do
-    dollar_and_cents =
-      cents
-      |> Decimal.div(100)
-      |> Decimal.round(2)
-      "$#{dollar_and_cents}"
-  end
 end
